refactor: improve test coverage
