+++ package.json
@@ -28,7 +28,7 @@
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
     "test": "vitest run",
     "test:watch": "vitest",
-    "test:cov": "npm run test -- --config ./vitest.config.e2e.ts --coverage",
+    "test:cov": "npm run test -- --config ./vitest.config.cov.ts",
     "test:debug": "npm run test -- --inspect --logHeapUsage",
     "test:e2e": "vitest run --config ./vitest.config.e2e.ts",
     "test:e2e:watch": "vitest --config ./vitest.config.e2e.ts"
