+++ src/domain/product/application/use-cases/create-product.ts
@@ -9,7 +9,7 @@ import { Product } from '../../enterprise/entities/product';
 import {
   Administrator,
   AdministratorProps,
-} from '../../enterprise/entities/responsable';
+} from '../../enterprise/entities/administrator';
 import { CategoriesRepository } from '../gateways/repositories/categories-repository';
 import { ProductsRepository } from '../gateways/repositories/products-repository';
 import { File } from '../gateways/storage/file';
