+++ src/domain/product/enterprise/entities/administrator.ts
